Improve the dev run-script
The index.html file is now generated by webpack, which will inject the
built scripts. The template is written in pug (I have also changed all
instances of jade to pug because of the rename).

The script injection uses the publicPath configuration variable; I have put
in a ternary expression that checks for the DEV environment variable. This
variable is set when using `npm run dev`, thus the publicPath will be set
accordingly (see webpack-dev-server).